(svn r1676) Increase the size of TileIndex and TileIndexDiff to 32bits and adapt the save/load data and some other parts of the code to that change
WARNING: If i made any mistake here it WILL lead to corrupted savegames!

+++ vehicle.h
@@ -102,8 +102,8 @@ struct Vehicle {
 	TileIndex tile;		// Current tile index
 	TileIndex dest_tile; // Heading for this tile
 
-	int16 x_pos;			// coordinates
-	int16 y_pos;
+	int32 x_pos;			// coordinates
+	int32 y_pos;
 	byte z_pos;
 	byte direction;		// facing
